remove non-integration iterate test and rename main files
